Removed dummy H1 space from API. Moved code for SampleMesh to cpp file.
